+++ Control/Joint/Concepts/Lens.hs
@@ -19,3 +19,9 @@ over lens f = extract . retrofit f . lens
 
 zoom :: Stateful bg t => Lens bg ls -> State ls a -> t a
 zoom lens (State f) = adapt . State $ (\(Store (p, g)) -> (\(x,y) -> (g x, y)) . f $ p) . lens
+
+_1 :: Lens (a, b) a
+_1 (x, y) = Store (x, \x' -> (x', y))
+
+_2 :: Lens (a, b) b
+_2 (x, y) = Store (y, \y' -> (x, y'))