+++ CHANGELOG.md
@@ -59,3 +59,4 @@
 
 # 0.1.9
 * Define `pos`, `seek`, `peek` and `retrofit` methods for `Store`
+* Define `Lens`, `view`, `set` and `over` in `Store` module